[GN-130] refactor: 메인페이지 리팩토링
