+++ src/app/(DefaultLayout)/(MainLayout)/page.tsx
@@ -1,150 +1,62 @@
 "use client";
 
 import Image from "next/image";
-import { useEffect, useState } from "react";
+import { useState } from "react";
 
 import ArrowButton from "./components/ArrowButton";
 import CategoryFilter from "./components/CategoryFilter";
-import activityService from "@/apis/activity/activity.service";
 import SortDropdown from "@/app/(DefaultLayout)/(MainLayout)/components/SortDropdown";
-import type { SortOptionType } from "@/app/(DefaultLayout)/(MainLayout)/components/SortDropdown";
 import Card from "@/components/card/Card";
 import Pagination from "@/components/pagination/Pagination";
 import { Search } from "@/components/search/Search";
-import type { ActivityCategoryType } from "@/types/activity";
+import { useActivities } from "@/hooks/useActivities";
+import { useAutoSlider } from "@/hooks/useAutoSlider";
+import { useCategoryFilter } from "@/hooks/useCategoryFilter";
+import { usePagination } from "@/hooks/usePagination";
+import { usePopularActivities } from "@/hooks/usePopularActivities";
+import { useResponsiveSlider } from "@/hooks/useResponsiveSlider";
+import { useSort } from "@/hooks/useSort";
 import { ActivityType } from "@/types/activity";
 
 function HomePage() {
-  const [activities, setActivities] = useState<ActivityType[]>([]);
-  const [popularActivities, setPopularActivities] = useState<ActivityType[]>(
-    [],
-  );
-  const [currentIndex, setCurrentIndex] = useState(0);
+  // 🔥 인기 체험 영역 관련 상태
+  const { popularActivities } = usePopularActivities();
   const [slideIndex, setSlideIndex] = useState(0);
-  const [selectedCategory, setSelectedCategory] = useState<
-    ActivityCategoryType | ""
-  >("");
-  const [currentPage, setCurrentPage] = useState(1);
-  const [totalPages, setTotalPages] = useState(0);
-  const [sortOption, setSortOption] = useState<SortOptionType>({
-    id: "latest",
-    title: "최신순",
-  });
 
-  const [itemsPerSlide, setItemsPerSlide] = useState(4);
-  const [itemsPerPage, setItemsPerPage] = useState(8);
+  // 🛼 모든 체험 영역 관련 상태
+  const { selectedCategory, handleCategorySelect } = useCategoryFilter();
+  const { currentPage, setCurrentPage } = usePagination();
+  const { sortOption, setSortOption } = useSort();
+  const { itemsPerSlide, itemsPerPage } = useResponsiveSlider();
 
-  useEffect(() => {
-    const updateItemsPerSlide = () => {
-      if (window.innerWidth < 1280) {
-        setItemsPerSlide(2);
-      } else {
-        setItemsPerSlide(4);
-      }
-    };
+  const { activities, totalPages } = useActivities(
+    selectedCategory,
+    sortOption.id,
+    currentPage,
+    itemsPerPage,
+  );
 
-    updateItemsPerSlide();
-    window.addEventListener("resize", updateItemsPerSlide);
-    return () => window.removeEventListener("resize", updateItemsPerSlide);
-  }, []);
-
-  useEffect(() => {
-    const updateItemsPerPage = () => {
-      if (window.innerWidth < 769) {
-        setItemsPerPage(6);
-      } else if (window.innerWidth < 1280) {
-        setItemsPerPage(4);
-      } else {
-        setItemsPerPage(8);
-      }
-    };
-
-    updateItemsPerPage();
-    window.addEventListener("resize", updateItemsPerPage);
-    return () => window.removeEventListener("resize", updateItemsPerPage);
-  }, []);
-
-  useEffect(() => {
-    const fetchActivities = async () => {
-      try {
-        const response = await activityService.getActivities({
-          query: {
-            method: "offset",
-            category: selectedCategory || undefined,
-            sort: sortOption.id,
-            page: currentPage,
-            size: itemsPerPage,
-          },
-        });
-        setActivities(response.data.activities);
-        setTotalPages(Math.ceil(response.data.totalCount / itemsPerPage));
-      } catch (error) {
-        console.error("Error fetching activities:", error);
-      }
-    };
-    fetchActivities();
-  }, [selectedCategory, currentPage, sortOption, itemsPerPage]);
-
-  useEffect(() => {
-    const fetchPopularActivities = async () => {
-      try {
-        const response = await activityService.getActivities({
-          query: {
-            method: "offset",
-            sort: "most_reviewed",
-          },
-        });
-        setPopularActivities(response.data.activities);
-      } catch (error) {
-        console.error("Error fetching popular activities:", error);
-      }
-    };
-    fetchPopularActivities();
-  }, []);
-
-  useEffect(() => {
-    if (activities.length === 0) return;
-
-    const interval = setInterval(() => {
-      setCurrentIndex((prevIndex) => {
-        if (prevIndex >= activities.length) {
-          return 1;
-        }
-        return prevIndex + 1;
-      });
-    }, 4000);
-
-    return () => clearInterval(interval);
-  }, [activities]);
-
-  useEffect(() => {
-    if (currentIndex === activities.length) {
-      const timeout = setTimeout(() => {
-        setCurrentIndex(0);
-        const slider = document.getElementById("slider-track");
-        if (slider) {
-          slider.style.transition = "none";
-          slider.style.transform = `translateX(0%)`;
-          // force reflow then restore transition
-          void slider.offsetWidth;
-          slider.style.transition = "";
-        }
-      }, 700); // matches duration-700
-
-      return () => clearTimeout(timeout);
-    }
-  }, [currentIndex, activities.length]);
-
-  const extendedActivities =
-    activities.length > 0 ? [...activities, activities[0]] : [];
+  // 🖼️ 배너 영역 관련 상태
+  const { currentIndex, extendedActivities } = useAutoSlider(activities);
 
   const paginatedActivities = popularActivities?.slice(
     slideIndex * itemsPerSlide,
     slideIndex * itemsPerSlide + itemsPerSlide,
   );
 
+  const renderCard = (activity: ActivityType) => (
+    <Card
+      key={activity.id}
+      title={activity.title}
+      price={activity.price}
+      bannerImageUrl={activity.bannerImageUrl}
+      rating={activity.rating}
+      reviewCount={activity.reviewCount}
+    />
+  );
+
   return (
-    <div className="flex flex-col gap-15   ">
+    <div className="flex flex-col space-y-16">
       {/* 배너 영역 */}
       <section className="w-full overflow-hidden relative">
         <div
@@ -158,7 +70,7 @@ function HomePage() {
           {extendedActivities.map((activity, index) => (
             <div
               key={`${activity.id}-${index}`}
-              className="w-full flex-shrink-0 mobile:h-[18.1rem] tablet:h-[37.5rem] desktop:h-[50rem] relative rounded-[24px] overflow-hidden"
+              className="w-full flex-shrink-0 relative rounded-[24px] overflow-hidden mobile:h-[18.1rem] tablet:h-[37.5rem] desktop:h-[50rem]"
               style={{ width: `${100 / (activities.length || 1)}%` }}
             >
               <Image
@@ -188,17 +100,8 @@ function HomePage() {
         <h2 className="text-24-b">🔥 인기 체험</h2>
         <div className="relative">
           {/* 데스크탑/태블릿용 그리드 */}
-          <div className="grid tablet:grid-cols-2 desktop:grid-cols-4 gap-4 hidden mobile:hidden tablet:grid">
-            {paginatedActivities?.map((activity) => (
-              <Card
-                key={activity.id}
-                title={activity.title}
-                price={activity.price}
-                bannerImageUrl={activity.bannerImageUrl}
-                rating={activity.rating}
-                reviewCount={activity.reviewCount}
-              />
-            ))}
+          <div className="hidden tablet:grid tablet:grid-cols-2 desktop:grid-cols-4 gap-4">
+            {paginatedActivities?.map((activity) => renderCard(activity))}
           </div>
 
           {/* 모바일용 가로 슬라이드 */}
@@ -208,13 +111,7 @@ function HomePage() {
                 key={activity.id}
                 className="flex-shrink-0 w-[85%] snap-start"
               >
-                <Card
-                  title={activity.title}
-                  price={activity.price}
-                  bannerImageUrl={activity.bannerImageUrl}
-                  rating={activity.rating}
-                  reviewCount={activity.reviewCount}
-                />
+                {renderCard(activity)}
               </div>
             ))}
           </div>
@@ -245,29 +142,11 @@ function HomePage() {
         <div className="flex overflow-x-auto gap-2 tablet:justify-between scrollbar-hide">
           <CategoryFilter
             selectedId={selectedCategory}
-            onSelect={(id: string) => {
-              const labelMap: Record<string, ActivityCategoryType> = {
-                culture: "문화 · 예술",
-                food: "식음료",
-                tour: "투어",
-                sightseeing: "관광",
-                wellbeing: "웰빙",
-              };
-              setSelectedCategory(labelMap[id]);
-            }}
+            onSelect={handleCategorySelect}
           />
         </div>
         <div className="grid grid-cols-2 tablet:grid-cols-2 desktop:grid-cols-4 gap-4">
-          {activities?.map((activity) => (
-            <Card
-              key={activity.id}
-              title={activity.title}
-              price={activity.price}
-              bannerImageUrl={activity.bannerImageUrl}
-              rating={activity.rating}
-              reviewCount={activity.reviewCount}
-            />
-          ))}
+          {activities?.map((activity) => renderCard(activity))}
         </div>
         <Pagination
           currentPage={currentPage}
