+++ src/hooks/usePagination.ts
@@ -1,0 +1,6 @@
+import { useState } from "react";
+
+export const usePagination = (initialPage = 1) => {
+  const [currentPage, setCurrentPage] = useState(initialPage);
+  return { currentPage, setCurrentPage };
+};
